+++ tests/index.test.ts
@@ -1,2 +1,3 @@
+import './access.test'
 import './openAlexMapper.test'
 import './zoteroResponse.test'